Add step to decompose connected components
It does not make sense to run an expensive
cut-matching step on a graph with several
connected components since the components define
clear and easy cuts. This change runs a connected
component check on a subgraph each time
'expander-decomp' is called. If multiple
components are found, recurse on them seperately.