Implement new graph datastructure 'SubsetGraph'
The purpose of this representation is to avoid
having to use hashsets to keep track of if a
vertex is 'alive' or not. The datastructure
supports two types of vertex remove operations
which are both restoreable. This is done by
reordering adjacency lists such that the first 'k'
edges are valid whenever a vertex is removed.