Fix slow matching algorithm
Matching is still O(n^2) but in practise it should
be fast on most graphs. Due to always resetting
the entire 'visited' vector the matching ran in
\Theta(n^2) time which is not good.

By changing the 'visited' vector from storing
booleans to storing integers we can define a
visited vertex 'u' as a vertex 'visited[u]'
matches the identifier the search was started
with.