Remove assertion for matching size
Assert was put in place to make sure all 'axLeft'
vertices were matched with some vertex in
'axRight'. If all flow is routed it is correct to
expect this. However, if not all flow is routed
and vertices are removed according to some level
cut it is no longer guaranteed that every vertex
can be matched using the same flow solution.